+++ runtime/klee-libc/mempcpy.c
@@ -9,7 +9,7 @@
 
 #include <stdlib.h>
 
-void *mempcpy(void *destaddr, void const *srcaddr, unsigned int len) {
+void *mempcpy(void *destaddr, void const *srcaddr, size_t len) {
   char *dest = destaddr;
   char const *src = srcaddr;
